refactor(ui): simplify timer screen composable
